Database Dumps
These are the Database Dumps for this project. These had to be exported with MySQLWorkbench 6.3.10 since there is currently a bug in the latest version preventing the user from creating dumps. Import should be alright, but if any errors are encountered, please revert your version of MySQLWorkbench or similar programs to compatible versions.